requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
python-multipart
pydantic-settings>=2.2.0,<3.0.0